requirements.txt
asgiref==3.8.1
Django==5.0.6
django-allauth==0.61.0
djangorestframework==3.15.1
dj-database-url==2.1.0
gunicorn==20.1.0
mysqlclient==2.2.4
python-dotenv==0.19.2
